CI: added test workflow CI: update action from setup-java@v2 to setup-java@v3 build: update java version fro 1.8 to 17 refactor: remove extra brackets in TeamDaoTest.kt refactor: add break in for loop to prevent unneeded looping
