+++ .github/workflows/test-release.yml
@@ -1,0 +1,27 @@
+name: Test Release
+
+on:
+  pull_request:
+    branches: [test_branch]
+
+jobs:
+  run_tests:
+    name: Testing release variant
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v3
+
+      - uses: actions/setup-java@v3
+        with:
+          distribution: 'temurin'
+          java-version: '17'
+
+      - name: Setup Gradle
+        uses: gradle/gradle-build-action@v2
+
+      - name: Make gradlew executable
+        run: chmod +x ./gradlew
+
+      - name: Execute Gradle command - testReleaseUnitTest
+        run: ./gradlew testReleaseUnitTest
