+++ app/build.gradle
@@ -41,11 +41,11 @@ android {
         //for supporting older APIs like(LocalDateTime, OffsetDateTime, DateTimeFormatter)
         //look at the implemented dependency
         coreLibraryDesugaringEnabled = true
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
     kotlinOptions {
-        jvmTarget = '1.8'
+        jvmTarget = '17'
     }
     buildFeatures {
         dataBinding true
