+++ app/src/androidTest/java/com/mobile/sportology/rooms/LeagueDaoTest.kt
@@ -1,14 +1,9 @@
-package com.mobile.sportology.roomsTest
+package com.mobile.sportology.rooms
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
-import androidx.room.Room
-import androidx.test.core.app.ApplicationProvider
-import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.SmallTest
 import com.google.common.truth.Truth.assertThat
 import com.mobile.sportology.models.football.LeagueRoom
-import com.mobile.sportology.rooms.AppDatabase
-import com.mobile.sportology.rooms.LeagueDao
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
 import kotlinx.coroutines.runBlocking
@@ -16,7 +11,6 @@ import org.junit.After
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import org.junit.runner.RunWith
 import javax.inject.Inject
 import javax.inject.Named
